+++ crates/types/src/lib.rs
@@ -18,15 +18,15 @@ mod mocks;
 mod query;
 mod response;
 mod result;
-mod serde;
+mod serializers;
 mod time;
 mod tx;
 mod utils;
 
 pub use {
     address::*, app::*, bank::*, binary::*, coin::*, context::*, db::*, empty::*, error::*,
-    event::*, hash::*, hashers::*, imports::*, math::*, query::*, response::*, result::*, serde::*,
-    time::*, tx::*, utils::*,
+    event::*, hash::*, hashers::*, imports::*, math::*, query::*, response::*, result::*,
+    serializers::*, time::*, tx::*, utils::*,
 };
 
 // Mocks need to be excluded in Wasm builds because they depend on k256/p256
